requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
pillow==10.0.0
pymongo==4.5.0
google-generativeai==0.3.0
requests==2.31.0
PyMuPDF==1.23.0
python-dotenv==1.0.0
logging==0.4.9.6
datetime
python-telegram-bot==20.4